feat(6746): extract command contract tests
