+++ crates/kamn-e2e-harness/tests/command_contract/external_execution_contract_tests.rs
@@ -1,0 +1,8 @@
+use super::*;
+
+#[path = "external_execution_contract_tests/runtime_execution_contract_tests.rs"]
+mod runtime_execution_contract_tests;
+#[path = "external_execution_contract_tests/filesystem_preflight_contract_tests.rs"]
+mod filesystem_preflight_contract_tests;
+#[path = "external_execution_contract_tests/config_flag_mapping_contract_tests.rs"]
+mod config_flag_mapping_contract_tests;
