+++ crates/kamn-e2e-harness/tests/command_contract/integration_runtime_contract_tests.rs
@@ -1,0 +1,8 @@
+use super::*;
+
+#[path = "integration_runtime_contract_tests/integration_config_contract_tests.rs"]
+mod integration_config_contract_tests;
+#[path = "integration_runtime_contract_tests/process_runtime_contract_tests.rs"]
+mod process_runtime_contract_tests;
+#[path = "integration_runtime_contract_tests/spawn_execution_contract_tests.rs"]
+mod spawn_execution_contract_tests;
