+++ crates/kamn-e2e-harness/tests/command_contract/phase_inventory_contract_tests.rs
@@ -1,0 +1,6 @@
+use super::*;
+
+#[path = "phase_inventory_contract_tests/phase_progression_contract_tests.rs"]
+mod phase_progression_contract_tests;
+#[path = "phase_inventory_contract_tests/lifecycle_summary_contract_tests.rs"]
+mod lifecycle_summary_contract_tests;
