+++ crates/kamn-e2e-harness/tests/command_contract/scenario_evidence_contract_tests.rs
@@ -1,0 +1,6 @@
+use super::*;
+
+#[path = "scenario_evidence_contract_tests/scenario_results_contract_tests.rs"]
+mod scenario_results_contract_tests;
+#[path = "scenario_evidence_contract_tests/evidence_contract_tests.rs"]
+mod evidence_contract_tests;
